apps/server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic
pydantic-settings==2.1.0
python-dotenv==1.0.0
openai
anthropic==0.7.8
mem0ai==0.0.12
httpx==0.25.2
python-multipart==0.0.6
mistralai
llama-api-client

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
bcrypt==4.1.1

# Encryption
cryptography==41.0.7